pi2/src/requirements.txt
flask
flask_cors
smbus2
RPi.bme280
RPi.GPIO
picamera
opencv-python
schedule